+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Version 0.7.1.5 (compiler 0.7.1.3, engine 0.7.3.4):
     to string. Was '\x00', is now '\0'.
   - Made MKR* instructions more memory efficient when explicit default values
     are encountered.
+  - Implemented the string formatting operator (str % tuple).
 
 Version 0.7.1.4 (compiler 0.7.1.2, engine 0.7.3.3):
 - Engine: